requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.6
django-adminlte-3==0.1.6
django-colorfield==0.8.0
django-import-export==3.1.0
django-mathfilters==1.0.0
django-pdfkit==0.3.1
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
pdfkit==0.6.0
Pillow==9.4.0
pytz==2022.7.1
PyYAML==6.0
sqlparse==0.4.3
tablib==3.3.0
wkhtmltopdf==0.2
xlrd==2.0.1
xlwt==1.3.0
